selling and equities computations

+++ database/migrations/2021_05_01_140949_create_equities_table.php
@@ -17,8 +17,8 @@ class CreateEquitiesTable extends Migration
             $table->id();
             $table->timestamps();
             $table->string('date'); 
-            $table->integer('total_equity');
-            $table->integer('remaining_cash');
+            $table->double('total_equity');
+            $table->double('remaining_cash');
             $table->string('action');
             $table->integer('action_reference_id');
         });
